+++ Sources/SwiftUIKit/Extensions/View+Screenshot.swift
@@ -1,0 +1,34 @@
+//
+//  View+Screenshot.swift
+//  SwiftUIKit
+//
+//  Created by Daniel Saidi on 2020-09-17.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+#if os(iOS)
+import SwiftUI
+
+extension View {
+    func takeScreenshot(origin: CGPoint, size: CGSize) -> UIImage {
+        let window = UIWindow(frame: CGRect(origin: origin, size: size))
+        let hosting = UIHostingController(rootView: self)
+        hosting.view.frame = window.frame
+        window.addSubview(hosting.view)
+        window.makeKeyAndVisible()
+        return hosting.view.renderedImage
+    }
+}
+
+private extension UIView {
+    
+    var renderedImage: UIImage {
+        UIGraphicsBeginImageContextWithOptions(bounds.size, false, 0.0)
+        let context = UIGraphicsGetCurrentContext()!
+        layer.render(in: context)
+        let image = UIGraphicsGetImageFromCurrentImageContext()!
+        UIGraphicsEndImageContext()
+        return image
+    }
+}
+#endif